Fucked up the launch pad.
Currently working on separating the roles of the book keeper and the book itself. Keeper will handle saving market book data to cache and the book itself will allow others to read from it.

It's starting to get late though and my mind is turning to mush. Time to put this down before I mess it all up.